% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate-grades.R
\name{apply_clobber}
\alias{apply_clobber}
\title{Apply Clobber Policy}
\usage{
apply_clobber(gs, policy)
}
\arguments{
\item{gs}{A vector of assignment scores and weights coming from a row of the
gs data frame.}

\item{policy}{A single-layer list containing, at least \code{assignments},
a vector and optionally \code{n_drops}, an integer.}
}
\value{
A gs dataframe with clobber policies applied to relevant categories
}
\description{
This computes any clobber policy, where the category grade is determined by the
maximum score between two categories.
}
