% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slip-days.R
\name{apply_slip_days}
\alias{apply_slip_days}
\title{Apply Slip Days}
\usage{
apply_slip_days(gs, policy)
}
\arguments{
\item{gs}{A Gradescope dataframe}

\item{policy}{A grading policy file as an R list}
}
\value{
A dataframe of the original Gradescope data with lateness updated according to the slip day policies
}
\description{
This function applies any slip day policies within the policy file.
}
\examples{
apply_slip_days(gs = gs_demo, policy = slip_days_policy_demo)

}
