% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assignment.R
\name{assignment}
\alias{assignment}
\title{Create an assignment}
\usage{
assignment(
  course_id,
  assignment_id,
  assignment_category,
  out_of,
  due_date,
  rubric = NULL,
  portal = NULL,
  overwrite = TRUE
)
}
\arguments{
\item{course_id}{desc}

\item{assignment_id}{desc}

\item{assignment_category}{desc}

\item{out_of}{desc}

\item{due_date}{desc}

\item{rubric}{desc}

\item{portal}{desc}

\item{overwrite}{desc}
}
\value{

}
\description{
Create an assignment
}
