% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assignments_gs.R
\name{assignments_gs}
\alias{assignments_gs}
\title{Create Assignments Table from Input Data}
\usage{
assignments_gs(unique_student_ids)
}
\arguments{
\item{assignments_gs}{A dataframe (csv) from gradescope containing the unique student IDs information
(this is the output of the \code{unique_student_ids} function).}
}
\value{
A tibble dataframe with the processed assignment table.
}
\description{
This function takes the gradescope dataframe after some assignments have been created generating an assignment table.
}
