% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate-grades2.R
\name{by_max}
\alias{by_max}
\title{Calculate Grade by Maximum Score in Category}
\usage{
by_max(gs_data, sub_category, sub_assigns)
}
\arguments{
\item{gs_data}{a wide formatted, processed Gradescope data}

\item{sub_category}{name of sub-category grade that is being calculated}

\item{sub_assigns}{assignments within this subcategory}
}
\value{
a wide formatted dataframe with raw score, max points and percent for new subcategory
}
\description{
This functions calculates the grade of a category with the maximum percentage score of the category.
The max points of this new category is given by the average of the max points of assignments within
this category.
}
