% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate-grades2.R
\name{by_none}
\alias{by_none}
\title{No Grading Policy Function}
\usage{
by_none(gs_data)
}
\arguments{
\item{gs_data}{a wide formatted, processed Gradescope data}
}
\value{
the same wide formatted dataframe
}
\description{
This functions is used as a placeholder function when there is no grading policy.
}
