% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate-grades2.R
\name{by_percent}
\alias{by_percent}
\title{Calculate Grade Weighted by Percent}
\usage{
by_percent(gs_data, aggregate_assign, sub_assigns)
}
\arguments{
\item{gs_data}{a wide formatted, processed Gradescope data}

\item{sub_assigns}{assignments within this subcategory}

\item{sub_category}{name of sub-category grade that is being calculated}
}
\value{
a wide formatted dataframe with raw score, max points and percent for new subcategory
}
\description{
This functions calculates the grade of a category with each assignment weighted equally
}
