% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate-grades2.R
\name{by_points}
\alias{by_points}
\title{Calculate Grade Weighted by Points}
\usage{
by_points(gs_data, sub_category, sub_assigns)
}
\arguments{
\item{gs_data}{a wide formatted, processed Gradescope data}

\item{sub_category}{name of sub-category grade that is being calculated}

\item{sub_assigns}{assignments within this subcategory}
}
\value{
a wide formatted dataframe with raw score, max points and percent for new subcategory
}
\description{
This functions calculates the grade of a category with each assignment weighted by
how many points it is worth.
}
