% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate-grades2.R
\name{by_weight}
\alias{by_weight}
\title{Calculate Grade by Maximum Score in Category}
\usage{
by_weight(gs_data, sub_category, sub_assigns, weights)
}
\arguments{
\item{gs_data}{a wide formatted, processed Gradescope data}

\item{sub_category}{name of sub-category grade that is being calculated}

\item{sub_assigns}{assignments within this subcategory}

\item{weights}{weights in same order as \code{sub_assigns}}
}
\value{
a wide formatted dataframe with raw score, max points and percent for new subcategory
}
\description{
This functions calculates the grade of a category by weighting assignments by
given weights.
}
