% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate-grades2.R
\name{calculate_grades_with_for_loop}
\alias{calculate_grades_with_for_loop}
\title{Calculate Grades with For Loop}
\usage{
calculate_grades_with_for_loop(gs_data, flat_policy)
}
\arguments{
\item{gs_data}{processed and prepped Gradescope data}

\item{policy}{flattened and processed policy file}
}
\value{
the gs_data with appended subcategory and overall grades
}
\description{
This function calculates grades with the by-functions using a for loop
}
