% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lateness-in-pivot.R
\name{calculate_lateness}
\alias{calculate_lateness}
\title{Calculate Scores after Lateness
This function uses the merged pivotted data + lateness table to compute
raw scores after lateness.}
\usage{
calculate_lateness(lateness_table)
}
\arguments{
\item{lateness_table}{A processed pivotted data from Gradescope}
}
\value{
A data frame
}
\description{
Calculate Scores after Lateness
This function uses the merged pivotted data + lateness table to compute
raw scores after lateness.
}
