% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate-policy.R
\name{check_assignment_names}
\alias{check_assignment_names}
\title{Check Category Names}
\usage{
check_assignment_names(policy, pivot_df)
}
\arguments{
\item{policy}{policy file to validate}

\item{pivot_df}{pivotted gradescope data}
}
\value{
None
}
\description{
This functions checks that all assignments within categories of policy file exists in the Gradescope data
}
