% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate-policy.R
\name{check_cat_names}
\alias{check_cat_names}
\title{Check Category Names}
\usage{
check_cat_names(category_name)
}
\arguments{
\item{category_name}{the names within a category in a policy file}
}
\value{
returns logical vector
}
\description{
This functions checks the names of a single category in a policy file
}
