% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate-gs.R
\name{check_data_colnames_format}
\alias{check_data_colnames_format}
\title{Check Formatting of Column Names for Gradescope Data}
\usage{
check_data_colnames_format(gs_data)
}
\arguments{
\item{gs_data}{Grades data that's validated for having the right format in terms of names and nesting}
}
\value{
Outputs same dataframe if no error
}
\description{
This functions checks the column names throughout the Gradescope data
}
