% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate-gs.R
\name{check_data_format}
\alias{check_data_format}
\title{Check Formatting of Gradescope Data}
\usage{
check_data_format(gs_data)
}
\arguments{
\item{gs_data}{Gradescope data}
}
\value{
Same dataframe if no error
}
\description{
This functions checks the column names throughout the Gradescope data.
There must be an SID column and at least one assignment.
It also gives an alert for what id cols and assignments are in the data.
}
