% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate-gs.R
\name{check_data_names}
\alias{check_data_names}
\title{Check Names for Gradescope Data}
\usage{
check_data_names(gs_data)
}
\arguments{
\item{gs_data}{Grades data that's validated for having the right format in terms of names and nesting}
}
\value{
No output, only stops and warnings
}
\description{
This functions checks the names throughout the Gradescope data
}
