% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate-policy.R
\name{check_names}
\alias{check_names}
\title{Check Names for Policy File}
\usage{
check_names(policy)
}
\arguments{
\item{policy_file}{A policy file that's validated for having the right format in terms of names and nesting}
}
\value{
No output, only stops and warnings
}
\description{
This functions checks the names throughout the policy file
}
