% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate-policy.R
\name{check_values}
\alias{check_values}
\title{Check Values for Policy File}
\usage{
check_values(policy_file)
}
\arguments{
\item{policy_file}{A policy file that's validated for having the right format in terms of values and datatypes}
}
\value{
No output, only stops and warnings
}
\description{
This functions checks the values throughout the policy file
}
