% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate-grades2.R
\name{choose_aggregation}
\alias{choose_aggregation}
\title{Choose Aggregation Function}
\usage{
choose_aggregation(
  aggregation,
  gs_data,
  sub_category,
  sub_assigns,
  weights = c()
)
}
\arguments{
\item{aggregation}{a string denoting which aggregation to use}

\item{...}{other values needed for the relevant grade calculation}
}
\value{
the  wide formatted dataframe with the new calculated subcategory
}
\description{
This function determines which aggregation function to apply for calculating
the grade of a subcategory.
}
