% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complex_syllabus_data.R
\docType{data}
\name{complex_syllabus_data}
\alias{complex_syllabus_data}
\title{Complex Syllabus Data}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 5 rows and 36 columns.
}
\usage{
complex_syllabus_data
}
\description{
Sample grades data for a complex syllabus, shown in "Case Study: A Complex Syllabus" data
}
\examples{
complex_syllabus_data

# yaml:as_yaml(complex_syllabus_data)

}
\keyword{datasets}
