% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lateness-in-pivot.R
\name{compute_lateness}
\alias{compute_lateness}
\title{Compute Lateness
This function computes scores after lateness policy has been applied.}
\usage{
compute_lateness(wide_processed_data, flat_policy)
}
\arguments{
\item{wide_processed_data}{A processed data in wide format}

\item{flat_policy}{A flattened policy file}
}
\value{
A data frame
}
\description{
Compute Lateness
This function computes scores after lateness policy has been applied.
}
