% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate-grades2.R
\name{compute_percent}
\alias{compute_percent}
\title{Compute Percentages}
\usage{
compute_percent(processed_gs_data)
}
\arguments{
\item{processed_gs_data}{a wide formatted, processed Gradescope data}
}
\value{
a wide formatted dataframe with percentages
}
\description{
This function computes percents by dividng the raw_scores by max_points for
each assignment in a wide formatted, processed Gradescope data
}
