% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate-grades.R
\name{convert_to_min}
\alias{convert_to_min}
\title{Convert to Min}
\usage{
convert_to_min(hms)
}
\arguments{
\item{hms}{time in HH:MM:SS format}
}
\value{
number of minutes
}
\description{
Converted a time in HH:MM:SS format to minutes
}
