% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate-policy.R
\name{copy_element_to_children}
\alias{copy_element_to_children}
\title{Copy list element to child assignment categories}
\usage{
copy_element_to_children(category, key)
}
\arguments{
\item{category}{A list from a policy file corresponding to a category like "Labs".
Must contain an element called \code{assignments}.}

\item{key}{A character string of the name of the element that you wish to copy.}
}
\value{
A list of the same structure as the input category, but with specified
element copied to all child categories that lack an element of that name.
}
\description{
A recursive function to copy an element of a policy file to
all children categories that lack that element. If a child category has that
element, that existing child element will not get overwritten by the parent element.
}
\details{
Can be used to propagate a field like \code{lateness} to all child categories of
a given category.
}
