% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_files.R
\name{create_assigns_table}
\alias{create_assigns_table}
\title{Create an Assignments Table from Policy File
This function creates an assignment table where each row is an assignment with its category and relevant category information.
This is an intermediary table used to combine the policy file and Gradescope data}
\usage{
create_assigns_table(policy)
}
\arguments{
\item{policy}{A validated policy file}
}
\value{
A data frame
}
\description{
Create an Assignments Table from Policy File
This function creates an assignment table where each row is an assignment with its category and relevant category information.
This is an intermediary table used to combine the policy file and Gradescope data
}
\examples{
# Example
create_assigns_table(policy_demo)
}
