% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cat_table.R
\name{create_empty_policy_file}
\alias{create_empty_policy_file}
\title{Create an Empty Policy File}
\usage{
create_empty_policy_file(num_cat = 1)
}
\arguments{
\item{num_cat}{number of categories in policy file}
}
\value{
An empty policy file as a list
}
\description{
This function creates an empty policy file with a certain number of categories.
}
\examples{
create_empty_policy_file(num_cat = 3)
}
