% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lateness-in-pivot.R
\name{create_lateness_table}
\alias{create_lateness_table}
\title{Create an Lateness Table
This function creates an assignment table where each row is an assignment with its category and relevant lateness information.}
\usage{
create_lateness_table(flat_policy)
}
\arguments{
\item{flat_policy}{A validated policy file in flattened form}
}
\value{
A data frame
}
\description{
Create an Lateness Table
This function creates an assignment table where each row is an assignment with its category and relevant lateness information.
}
