% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate-grades2.R
\name{create_overall_category}
\alias{create_overall_category}
\title{Create Overall Grade Category in Policy File}
\usage{
create_overall_category(policy_nested)
}
\arguments{
\item{policy_nested}{policy file with (nested) categories}
}
\value{
a policy file with an additional "Overall Grade" category
}
\description{
This function creates a super-category called "Overall Grade" within the policy file.
This represents the overall grade for the course.
This must be done BEFORE flattening policy file.
}
