% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_files.R
\name{data_merge_assigns}
\alias{data_merge_assigns}
\title{Merges processed pivot data
This function merges processed pivot data with assignment table}
\usage{
data_merge_assigns(assigns_table, pivot_df)
}
\arguments{
\item{assigns_table}{An assignment table made from the policy file}

\item{pivot_df}{A processed pivotted data from Gradescope}
}
\value{
A data frame
}
\description{
Merges processed pivot data
This function merges processed pivot data with assignment table
}
