% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{default_rubric}
\alias{default_rubric}
\title{Default rubric}
\format{
An object of class \code{list} of length 4.
}
\usage{
default_rubric
}
\description{
Template for a simple rubric to get you started in designing your own rubrics.
}
\keyword{datasets}
