% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cat_table.R
\name{delete_category}
\alias{delete_category}
\title{Deletes category with name "name"}
\usage{
delete_category(policy_list, name)
}
\arguments{
\item{policy_list}{a policy list if adding category to previous one.}

\item{name}{name of category.}
}
\value{
A list.
}
\description{
Deletes category with name "name"
}
