% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demo_data.R
\docType{data}
\name{demo_data}
\alias{demo_data}
\title{Demo Data}
\format{
\subsection{\code{demo_data}}{

A data frame with 10 rows and 44 columns:
\describe{
\item{Names}{student name}
\item{Email}{student email}
\item{SID}{student ID number}
\item{Sections}{student class section}
\item{Lab 1}{assignment called Lab 1}
\item{Lab 1 - Max Points}{maximum possible points for Lab 1}
\item{Lab 1 - Lateness (H:M:S)}{how late Lab 1 was turned in, formatted HH:MM:SS}
...
}
}
}
\source{
\url{https://www.who.int/teams/global-tuberculosis-programme/data}
}
\usage{
demo_data
}
\description{
A fake dataset that mimics Gradescope .csv format
}
\keyword{datasets}
