% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-gs.R
\name{determine_grade_source}
\alias{determine_grade_source}
\title{This function performs the auto-determination of where the grades dataframe was sourced.}
\usage{
determine_grade_source(grades_df)
}
\arguments{
\item{grades_df}{This is the input Dataframe}
}
\value{
A string "Canvas", "Gradescope", or "Unrecognized" of the determination.
}
\description{
This function performs the auto-determination of where the grades dataframe was sourced.
}
