% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drop_lowest_grade.R
\name{drop_lowest_grade}
\alias{drop_lowest_grade}
\title{Drop lowest grade (convert to exemption)}
\usage{
drop_lowest_grade(grades, assignment_categories)
}
\arguments{
\item{grades}{The result of \code{view_status}, filtered to have only assessments of interest (e.g., already graded and already due).}

\item{assignment_categories}{Character vector of \code{assigment_category}'s in which you want to drop the lowest grade.}
}
\value{
Revised grades \code{data.frame}, where the lowest grade in each category is changed to \code{exemption == TRUE} (no rows are removed).
}
\description{
Drop lowest grade (convert to exemption)
}
