% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate-grades.R
\name{drop_na_assignments}
\alias{drop_na_assignments}
\title{Drop NA Assignments}
\usage{
drop_na_assignments(merged_files)
}
\arguments{
\item{merged_files}{merged file of pivot_df + policy}
}
\value{
data frame with only assigned assignments
}
\description{
This function drops any assignments that aren't assigned to a category
}
