% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-gs.R
\name{drop_ungraded_assignments}
\alias{drop_ungraded_assignments}
\title{Drop Ungraded Assignments}
\usage{
drop_ungraded_assignments(gs_data, give_alert = TRUE)
}
\arguments{
\item{gs_data}{A Gradescope data}

\item{give_alert}{whether or not to return an alert of assignments}
}
\value{
same dataframe without graded assignments
}
\description{
This functions drops any assignments that have no grades for any students and replaced -Inf values
}
