% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate-grades.R
\name{drops}
\alias{drops}
\title{Accounting for Drops}
\usage{
drops(merged_files)
}
\arguments{
\item{merged_files}{merged file with only assigned assignments}
}
\value{
data frame with drops accounted for
}
\description{
This function accounts the allowed drops per category and removes the dropped assignments
from further calculations
}
