% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/email_student_reports.R
\name{email_student_reports}
\alias{email_student_reports}
\title{Email semester reports to students}
\usage{
email_student_reports(
  course_id,
  student_id = NULL,
  report_date = lubridate::ymd(lubridate::date(Sys.time())),
  your_email = "ekezell@sewanee.edu",
  email_body =
    "Attached you will find a report indicating your overall grade in the course to date, with a summary of each assignment grade I have on file for you. Please review this and reply to me if you believe any of my records are erroneous. Note that any overdue assignment (ones currently with a zero) can be updated if you submit the make-up materials.\\n\\nPlease let me know if you have any questions or concerns.\\n\\nBest wishes!\\n\\nProf. Ezell",
  json_path = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{course_id}{Course ID.}

\item{student_id}{Optionally specify a student; otherwise unsent grades for all students will be sent.}

\item{report_date}{The date of the report to share, provided as a character string (format yyyy-mm-dd). Default is the current date according to your system's time.}

\item{your_email}{Your email address.}

\item{email_body}{The main body of the email message; this will appear one line below the greeting to the student.}

\item{json_path}{File path to your \code{json} credentials on your machine. This is necessary to send emails via the package \code{gmailr}.}

\item{verbose}{Print updates to console?}
}
\value{
desc
}
\description{
Email semester reports to students
}
