% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate-grades.R
\name{equally_weighted}
\alias{equally_weighted}
\title{Aggregation Functions}
\usage{
equally_weighted(grades_mat, category, assignments)
}
\arguments{
\item{grades_mat}{Matrix with assignments + associated cols for that category}

\item{category}{Category name}

\item{assignments}{Assignment names for this category
A collection of functions that calculate score for assignments}
}
\description{
\itemize{
\item \code{equally_weighted()} takes the mean of scores
\item \code{none()} returns a single assignment score as category score;
if more than 1 assignment, defaults to equally weighted approach
}
}
