% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate-policy.R
\name{extract_nested}
\alias{extract_nested}
\title{Unnest subcategories from a category}
\usage{
extract_nested(category)
}
\arguments{
\item{category}{An element of a policy file, itself an R list.}
}
\value{
A list of categories (themselves lists of depth 1) beginning the with the
most deeply nested and ending with the outer layer of the input category.
}
\description{
Takes a top level category of a policy file (e.g. problem sets) and acts recursively
to extract all categories that are nested within and return them as a list of a
single level.
}
\examples{
# Example
extract_nested(policy_demo[[1]])
}
