% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find-weights.R
\name{find_weights}
\alias{find_weights}
\title{Find Weights for Weighted Mean Categories}
\usage{
find_weights(policy)
}
\arguments{
\item{policy}{Un-flattened Policy File}
}
\description{
This function will return a policy file where the weights have been extracted from
assignments into the above category featuring aggregation "weighted_mean", where the returned policy file is
otherwise identical to the input.
}
