% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process-policy.R
\name{flatten_policy}
\alias{flatten_policy}
\title{Flatten Policy File ("Un-nest" the Nested Structure of Policy File)}
\usage{
flatten_policy(policy)
}
\arguments{
\item{policy}{A valid policy file stored as a list.}
}
\value{
A single level list of all categories and subcategories ordered where
all leaves will precede the parent category in the list order.
}
\description{
This function reshape policy file from nested to flat structure by cycling
through the top-level categories of a policy file and un-nests all
subcategories to create a single level list of all categories and
subcategories.
}
\examples{
# Example
flatten_policy(policy_demo)
}
