% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process-gs.R
\name{get_assignments}
\alias{get_assignments}
\title{Get the Assignment Names for Unprocessed Gradescope Data}
\usage{
get_assignments(gs_data, give_alert = TRUE)
}
\arguments{
\item{gs_data}{unprocessed Gradescope dataframe}

\item{give_alert}{whether or not to return an alert of assignments}
}
\value{
vector
}
\description{
This function identified the assignments from Gradescope data
}
