% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process-gs.R
\name{get_assignments_unprocessed_data}
\alias{get_assignments_unprocessed_data}
\title{Get the Assignment Names for Unprocessed Gradescope Data}
\usage{
get_assignments_unprocessed_data(df, give_alert = TRUE)
}
\arguments{
\item{gs_data}{unprocessed Gradescope dataframe}
}
\value{
a list of assignments
}
\description{
This function identified the assignments from unprocessed gradescope data
}
