% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cat_table.R
\name{get_cat_index}
\alias{get_cat_index}
\title{Finds index of category within a policy list}
\usage{
get_cat_index(policy_list, name)
}
\arguments{
\item{policy_list}{a policy list if adding category to previous one.}

\item{name}{name of category.}
}
\value{
An integer.
}
\description{
Finds index of category within a policy list
}
\examples{
get_cat_index( update_category(name = "LABS"), "LABS")
}
