% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-policy.R
\name{get_categories}
\alias{get_categories}
\title{Get the Category Names for Policy}
\usage{
get_categories(policy, verbose = FALSE)
}
\arguments{
\item{policy}{R list of a valid policy file}

\item{verbose}{Whether or not to return an alert of categories}
}
\value{
A vector of the names of the categories in the policy R list
}
\description{
This function identifies the names of all categories in a policy list.
}
\examples{
get_categories(policy_demo, verbose = TRUE)

}
