% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate-grades.R
\name{get_category_grades}
\alias{get_category_grades}
\title{Get category grades for all students}
\usage{
get_category_grades(gs, policy)
}
\arguments{
\item{gs}{A vector of assignment scores and weights coming from a row of the
gs data frame.}

\item{policy_item}{A single-layer list containing, at least \code{assignments},
a vector and optionally \code{n_drops}, an integer.}
}
\value{
An extended version of the gs data frame, with columns added for each
category described in the policy file containing the grades for each student.
}
\description{
Iterates \code{get_one_grade()} across all categories of a policy file and across
all students present in a gs file.
}
