% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process-gs.R
\name{get_duplicate_ids}
\alias{get_duplicate_ids}
\title{Get Duplicate Student IDs}
\usage{
get_duplicate_ids(gs_data)
}
\arguments{
\item{gs_data}{Gradescope data containing a column named "SID" which holds student IDs.}
}
\value{
A dataframe "get_duplicate_ids": A dataframe containing all duplicate students from the data.
}
\description{
Returns all the duplicate Student IDs in the data.
}
\examples{
# Example dataframe
data <- data.frame(
  SID = c(3032412514, NA, 3032412516, 3032412517, 
          3032412518, 3032412519, 3032412520, 3032412521, 3032412521),
  names = c("John Smith", "Jane Doe", "Robert Brown", "Emily Johnson",
           "Michael Davis", "Linda Wilson", "James Taylor", 
           "Patricia Anderson", "Patricia Anderson"),
  email = c("john.smith@berkeley.edu", "jane.doe@berkeley.edu", 
  "robert.brown@berkeley.edu","emily.johnson@berkeley.edu",
   "michael.davis@berkeley.edu","linda.wilson@berkeley.edu",
   "james.taylor@berkeley.edu",
   "patricia.anderson@berkeley.edu", "patricia.anderson@berkeley.edu"
            )
 )
processed_data <- get_duplicate_ids(data)

}
