% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process-gs.R
\name{get_id_cols}
\alias{get_id_cols}
\title{Get the ID Columns for Gradescope Data}
\usage{
get_id_cols(df)
}
\arguments{
\item{gs_data}{Gradescope dataframe}
}
\value{
a list of id columns
}
\description{
This function identified the id columns from gradescope data
}
\examples{
df <- tibble::tibble(
name = 'name',
email = 'email',
sid = '343432',
`PS 1_-_Raw_Score` = 1,
`PS 1_-_Max_points` = 1,
`PS 1_-_Submission_time` = '1/21/2023 10:00:00 AM',
`PS 1-_-Lateness_(h_m_s)` = '0:00:00'
)

}
