% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process-gs.R
\name{get_id_cols_unprocessed_data}
\alias{get_id_cols_unprocessed_data}
\title{Get the ID Columns for Unprocessed Gradescope Data}
\usage{
get_id_cols_unprocessed_data(df, give_alert = TRUE)
}
\arguments{
\item{gs_data}{unprocessed Gradescope dataframe}
}
\value{
a list of unprocessed id columns
}
\description{
This function identified the id columns from unprocessed gradescope data
}
