% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate-grades.R
\name{get_letter_grades}
\alias{get_letter_grades}
\title{Get letter grades for all students}
\usage{
get_letter_grades(gs, policy)
}
\arguments{
\item{gs}{A vector of assignment scores and weights coming from a row of the
gs data frame with all category and overall grades}

\item{policy}{A policy file with course description, grading policy and
bounds for letter grades}
}
\value{
An extended version of the gs data frame, with a column added for
letter grades, allotted as described in the policy file
}
\description{
Get letter grades for all students
}
