% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-policy.R
\name{get_nested_assignments}
\alias{get_nested_assignments}
\title{Get the Nested Assignments for Category}
\usage{
get_nested_assignments(category, verbose = FALSE)
}
\arguments{
\item{category}{a category from a policy R list}

\item{verbose}{Whether or not to return an alert of assignments}
}
\value{
A vector of the names of the nested assignments in the category
}
\description{
This function identifies the nested assignments for a given category.
}
\examples{
category <- policy_demo[["categories"]][[1]]
get_nested_assignments(category = category, verbose = TRUE)

}
