% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate-grades.R
\name{get_one_grade}
\alias{get_one_grade}
\title{Get one category grade}
\usage{
get_one_grade(gs_row, policy_item)
}
\arguments{
\item{gs_row}{A vector of assignment scores and weights coming from a row of the
gs data frame.}

\item{policy_item}{A single-layer list containing, at least \code{assignments},
a vector and optionally \code{n_drops}, an integer.}
}
\value{
A single aggregated category grade.
}
\description{
Applies one of the aggregation functions to grade data from one student to
calculate one category grade for that student.
}
