% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grade.R
\name{grade}
\alias{grade}
\title{Launch grading app}
\usage{
grade(
  penalty_choices = c("N/A", "Late", "Behavior", "Other"),
  greeting =
    "Dear STUDENT,\\n\\nWell-done here. I particularly appreciate \\n\\nMoving forward, I suggest focusing primarily upon \\n",
  conclusion = "\\n\\nThank you again, STUDENT, for your hard work,\\nProf. Ezell",
  canned_comments = NULL,
  scroll_height = 200,
  render = TRUE
)
}
\arguments{
\item{penalty_choices}{Character vector of options for explaining a grade penalty, if assigned.}

\item{greeting}{Character vector}

\item{conclusion}{Character vector}

\item{canned_comments}{Optional URL (character vector) to a GoogleSheet with canned comments. Two columns in sheet: 'category' and 'comment'}

\item{scroll_height}{Default height of student scrolling area, in pixels.}

\item{render}{Default \code{TRUE} to render each grade's report as a \code{.PDF} as you work.}
}
\value{
This function launches a \code{Shiny} app that lets you grade submissions.
}
\description{
Launch grading app
}
