% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{grade_scale1}
\alias{grade_scale1}
\title{Grade scale 1}
\format{
An object of class \code{data.frame} with 5 rows and 4 columns.
}
\usage{
grade_scale1
}
\description{
Template for a classic 5-tier grade scale
}
\keyword{datasets}
