% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{grade_scale3}
\alias{grade_scale3}
\title{Grade scale 3}
\format{
An object of class \code{data.frame} with 3 rows and 4 columns.
}
\usage{
grade_scale3
}
\description{
Template for a simple 3-tier grade scale for simple adherence to instructions.
}
\keyword{datasets}
