% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gs_demo.R
\docType{data}
\name{gs_demo}
\alias{gs_demo}
\title{Gradescope Demo Data}
\format{
\subsection{\code{gs_demo}}{

A data frame with 10 rows and 44 columns:
\describe{
\item{Names}{student name}
\item{Email}{student email}
\item{SID}{student ID number}
\item{Sections}{student class section}
\item{Lab 1}{assignment called Lab 1}
\item{Lab 1 - Max Points}{maximum possible points for Lab 1}
\item{Lab 1 - Lateness (H:M:S)}{how late Lab 1 was turned in, formatted HH:MM:SS}
\item{Lab 1 - Submission Time}{when Lab 1 was turned in, formatted MM/DD/YY HH:MM}
...
}
}
}
\usage{
gs_demo
}
\description{
An example dataset that mimics Gradescope .csv format  with students as rows
and assignment information across the columns..
All following assignments follow same four-column format of Lab 1, as noted below.
assignments in gs_demo are as follows: Lab 2, Lab 3, Lab 4, Lab 5, Lab 6, Discussion 1, Discussion 2, Midterm and Final.
Any other columns that do not follow this format are designated as ID columns.
}
\keyword{datasets}
