% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate-grades.R
\name{intermediary_computations}
\alias{intermediary_computations}
\title{Intermediary Grading Computations}
\usage{
intermediary_computations(merged_files)
}
\arguments{
\item{merged_files}{merged file with only assigned assignments}
}
\value{
data frame with only assigned assignments and intermediary columns
}
\description{
This function creates columns of computations needed for grading, like number of relevant
assignments and late_time1/late_time2 in terms of minutes
}
