% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate-grades.R
\name{lateness}
\alias{lateness}
\title{Calculating Lateness}
\usage{
lateness(merged_files)
}
\arguments{
\item{lateness}{merged file with only assigned assignments}
}
\value{
data frame with only assigned assignments including lateness accounted in score
}
\description{
This function calculates lateness; adds a column for calculated score after lateness is accounted for
}
