% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lateness-in-pivot.R
\name{lateness_scores_in_wide}
\alias{lateness_scores_in_wide}
\title{Lateness Scores in Wide Format
This function pivots lateness scores from tall format to wide format.}
\usage{
lateness_scores_in_wide(pivotted_late_scores)
}
\arguments{
\item{pivotted_late_scores}{A pivotted dataframe with scores after lateness}
}
\value{
A data frame
}
\description{
Lateness Scores in Wide Format
This function pivots lateness scores from tall format to wide format.
}
