% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate-gs.R
\name{lower_colnames}
\alias{lower_colnames}
\title{Check Column Names for Gradescope Data}
\usage{
lower_colnames(processed_data)
}
\arguments{
\item{gs_data}{A dataframe with column names.}
}
\value{
same dataframe with lower case column names
}
\description{
This functions changes to lower case all column names throughout the Gradescope data
}
\examples{
data <- data.frame(
  'SID' <- c(123,345),
  'NAME' = c('Al', 'Ben'),
  'EMAIL' = c('al@company.com', 'ben@company.com')
)
lower <- lower_colnames(data)

}
