% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-grades.R
\name{mean_lateness}
\alias{mean_lateness}
\alias{sum_lateness}
\alias{max_lateness}
\title{Aggregation for Lateness Functions}
\usage{
mean_lateness(grades_mat, category, assignments)

sum_lateness(grades_mat, category, assignments)

max_lateness(grades_mat, category, assignments)
}
\arguments{
\item{grades_mat}{Matrix with assignments + associated columns for each category}

\item{category}{Category name}

\item{assignments}{Assignment names for this category}
}
\value{
A matrix with the aggregated lateness for each category
}
\description{
A collection of functions to computes lateness for category.
\itemize{
\item \code{mean_lateness()} computes the aggregated lateness of a category by taking the mean of the lateness of the category's assignments
\item \code{sum_lateness()} computes the aggregated lateness of a category by taking the sum of the lateness of the category's assignments
\item \code{max_lateness()} computes the aggregated lateness of a category by taking the maximum of the lateness of the category's assignments
}
}
\concept{{Aggregation for Lateness Functions}}
