% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moderate_syllabus_data.R
\docType{data}
\name{moderate_syllabus_data}
\alias{moderate_syllabus_data}
\title{Moderate Syllabus Data}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 5 rows and 32 columns.
}
\usage{
moderate_syllabus_data
}
\description{
Sample grades data for a moderately complex syllabus, shown in "Case Study: A Moderate Syllabus" data
}
\examples{
moderate_syllabus_data

# yaml:as_yaml(moderate_syllabus_data)

}
\keyword{datasets}
