% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lateness-in-pivot.R
\name{pivot_for_lateness}
\alias{pivot_for_lateness}
\title{Pivot for Lateness
This function merges processed pivot data with lateness table}
\usage{
pivot_for_lateness(lateness_table, pivot_df)
}
\arguments{
\item{pivot_df}{A processed pivotted data from Gradescope}

\item{assigns_table}{A lateness table made from the policy file}
}
\value{
A data frame
}
\description{
Pivot for Lateness
This function merges processed pivot data with lateness table
}
