% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process-gs.R
\name{pivot_gs}
\alias{pivot_gs}
\title{Convert GS data to longer table}
\usage{
pivot_gs(gs_data, names_sep = " - ")
}
\arguments{
\item{gs_data}{A dataframe (csv from Gradescope) containing a column named "sid" which holds student IDs.}

\item{names_sep}{String that controls how names are split during the pivot}
}
\value{
A dataframe "student_assignments_long": A dataframe with a unit of observation a student + each assignment.
}
\description{
This function takes a gradescope dataframe after it has been processed
to remove duplicate students or invalid IDs
and it applies pivot longer function to allow for unit of observation be a student + individual assignment
}
