% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_assignment_calendar.R
\name{plot_assignment_calendar}
\alias{plot_assignment_calendar}
\title{Plot cumulative points throughout semester}
\usage{
plot_assignment_calendar(course_id)
}
\arguments{
\item{course_id}{}
}
\value{
A \code{ggplot}.
}
\description{
Plot cumulative points throughout semester
}
