% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_assignment_weights.R
\name{plot_assignment_weights}
\alias{plot_assignment_weights}
\title{Plot assignment weights}
\usage{
plot_assignment_weights(course_id)
}
\arguments{
\item{course_id}{Course ID}
}
\value{
A \code{ggplot}
}
\description{
Plot assignment weights
}
