% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate-grades2.R
\name{prep_for_grading}
\alias{prep_for_grading}
\title{Prep Processed Gradescope}
\usage{
prep_for_grading(processed_gs_data, wide = TRUE)
}
\arguments{
\item{processed_gs_data}{a wide formatted, processed Gradescope data}

\item{wide}{TRUE if in wide format, FALSE if in pivotted format}
}
\value{
a wide formatted dataframe with scores
}
\description{
This function prepares processed Gradescope data for grading by computing percentages.
If wide == TRUE, this function removes all lateness and submission columns
}
