% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process-gs.R
\name{process_assignments}
\alias{process_assignments}
\title{Remove white spaces and special characters from column names}
\usage{
process_assignments(processed_data)
}
\arguments{
\item{processed_data}{A dataframe (csv from Gradescope)}
}
\value{
A dataframe "processed_data": Returns the same dataframe with modified column names.
}
\description{
This function takes a gradescope dataframe and
it removes white spaces and special characters from column names
}
\examples{

#gradebook gradescope demo data:
gs_demo
processed_assignments_df <- process_assignments(gs_demo)

}
