% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process-gs.R
\name{process_gs}
\alias{process_gs}
\title{Process Gradescope Data}
\usage{
process_gs(gs_data)
}
\arguments{
\item{gs_data}{A dataframe (csv from Gradescope)}
}
\value{
A dataframe
}
\description{
This functino merges any duplicates students and drops ungraded assignments
for faster processing.
}
