% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process-gs.R
\name{process_gs_data}
\alias{process_gs_data}
\title{Process Gradescope Data}
\usage{
process_gs_data(gs_data, wide_format = TRUE)
}
\arguments{
\item{gs_data}{A dataframe (csv directly from Gradescope)}

\item{wide_format}{if TRUE, keeps in wide format; if FALSE, pivots data}
}
\value{
Returns processed Gradescope data in wide or pivotted format
}
\description{
This function takes a gradescope dataframe and processes it completely.
}
