% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate-gs.R
\name{process_id}
\alias{process_id}
\title{Process Student IDs}
\usage{
process_id(gs_data)
}
\arguments{
\item{gs_data}{A dataframe (csv from Gradescope) containing a column named "sid" which holds student IDs.}
}
\value{
A dataframe "unique_ids": A dataframe containing unique student IDs.
}
\description{
This function processes a dataset with student IDs. It handles
erroneous student IDs by filtering out duplicates and handling NA values.
The function returns a dataframe with unique student IDs
}
\examples{
# Example dataframe
data <- data.frame(
  sid = c(3032412514, NA, 3032412516,
          3032412517, 3032412518, 3032412519, 3032412520, 3032412521, 3032412521),

  name = c("John Smith", "Jane Doe", "Robert Brown", "Emily Johnson",
           "Michael Davis", "Linda Wilson", "James Taylor", "Patricia Anderson", "Patricia Anderson"),
  email = c("john.smith@berkeley.edu", "jane.doe@berkeley.edu", "robert.brown@berkeley.edu",
            "emily.johnson@berkeley.edu", "michael.davis@berkeley.edu",
            "linda.wilson@berkeley.edu", "james.taylor@berkeley.edu",
            "patricia.anderson@berkeley.edu", "patricia.anderson@berkeley.edu"
            )
 )
processed_data <- process_id(data)

}
