% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process-policy.R
\name{process_policy}
\alias{process_policy}
\title{Process Policy File}
\usage{
process_policy(policy, verbose = FALSE)
}
\arguments{
\item{policy}{R list of a valid policy file}

\item{verbose}{Whether or not to print messages and warnings}
}
\value{
a (un-nested) flattened policy R list
}
\description{
This function processes the policy file in order to compute grades and flattens
the nested structure of the file.
}
\examples{
process_policy(policy_demo, verbose = TRUE)


}
