% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate-policy.R
\name{process_policy_names}
\alias{process_policy_names}
\title{Process Assignment and Category Names in Policy File}
\usage{
process_policy_names(policy)
}
\arguments{
\item{policy}{policy file}
}
\value{
policy file with processed names
}
\description{
This functions processes all assignment and subcategory names
by removing spaces, special characters, etc.
}
