% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_student_id.R
\name{process_student_id}
\alias{process_student_id}
\title{Process Student IDs}
\usage{
process_student_id(new_data)
}
\arguments{
\item{new_data}{A dataframe containing a column named "sid" which holds student IDs.}
}
\value{
A list with two dataframes:
\itemize{
\item "unique_sids": A dataframe containing unique student IDs.
\item "duplicates": A dataframe with rows having duplicate or NA student IDs.
}
}
\description{
This function processes a dataset with student IDs (sids). It handles
erroneous student IDs by filtering out duplicates and handling NA values.
The function returns a list containing two dataframes:
\enumerate{
\item "unique_sids" which has unique student IDs and
\item "duplicates" which contains all rows with duplicate or NA student IDs.
}
}
\examples{
# Example dataframe
data <- data.frame(
  sid = c(3032412514, NA, 3032412516,
          3032412517, 3032412518, 3032412519, 3032412520, 3032412521),

  name = c("John Smith", "Jane Doe", "Robert Brown", "Emily Johnson",
           "Michael Davis", "Linda Wilson", "James Taylor", "Patricia Anderson"),
  email = c("john.smith@berkeley.edu", "jane.doe@berkeley.edu", "robert.brown@berkeley.edu",
            "emily.johnson@berkeley.edu", "michael.davis@berkeley.edu",
            "linda.wilson@berkeley.edu", "james.taylor@berkeley.edu",
            "patricia.anderson@berkeley.edu"
            )
 )
processed_data <- process_student_id(data)

}
