% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-grades.R
\name{raw_over_max}
\alias{raw_over_max}
\title{Score Functions}
\usage{
raw_over_max(grades_mat, assignments)
}
\arguments{
\item{grades_mat}{Matrix with assignments + associated cols for that category}

\item{assignments}{Assignment names for this category}
}
\value{
A matrix with computed assignment scores
}
\description{
A collection of functions to calculate the percentage score and save back into raw-points column  (the columnn with the assignment score).
\itemize{
\item \code{raw_over_max()} computes score by dividing raw points by max points.
}
}
\concept{{Score Functions}}
