% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-gs.R
\name{read_canvas_grades}
\alias{read_canvas_grades}
\title{Takes in a dataframe uploaded from Canvas.
Returns a dataframe in our standardized format (ie like Gradescope).}
\usage{
read_canvas_grades(grades)
}
\description{
Takes in a dataframe uploaded from Canvas.
Returns a dataframe in our standardized format (ie like Gradescope).
}
