% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-gs.R
\name{read_files}
\alias{read_files}
\title{Read in 1 or more files from Grading Platform}
\usage{
read_files(grades_path, other_file_paths = list(), source = "auto")
}
\arguments{
\item{grades_path}{Is the path to the csv containing graded assignments.}

\item{other_file_paths}{Is a list containing any other filepaths desired to be read in. Each name should be a designated type of file. For example, "lateness" and "roster" and the value should be the filepath.}

\item{source}{Determines how to read the grade file. By default, it is auto so that Gradebook can determine what format the file is in. Alternatively, users can specify "Gradescope" or "Canvas" to force certain interpretations.}
}
\value{
A dataframe of grades. Later support for other files will be returned as a list as well.
}
\description{
This function will read in all input data to Gradebook.
}
