% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-gs.R
\name{read_gs}
\alias{read_gs}
\title{Read Gradescope CSV File}
\usage{
read_gs(path, verbose = FALSE)
}
\arguments{
\item{path}{Path to Gradescope CSV file}

\item{verbose}{Whether or not to print messages}
}
\value{
A dataframe of Gradescope CSV, if no errors
}
\description{
This functions reads the Gradescope CSV file and checks for correct format.
}
