% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-policy.R
\name{read_policy}
\alias{read_policy}
\title{Read Grading Policy .yml}
\usage{
read_policy(path, verbose = FALSE)
}
\arguments{
\item{path}{Path to policy file}

\item{verbose}{whether or not to print messages}
}
\value{
R list
}
\description{
This functions reads the grading policy YAML file, checks for correct format,
converts NA values to zeros and computes scores for each assignment.
}
