% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_class.R
\name{render_class}
\alias{render_class}
\title{Render grades for entire class}
\usage{
render_class(course_id)
}
\arguments{
\item{course_id}{Course ID}
}
\value{
A list.
}
\description{
Render grades for entire class
}
