% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_grade.R
\name{render_grade}
\alias{render_grade}
\title{Render the report for a grade}
\usage{
render_grade(grade_file)
}
\arguments{
\item{grade_file}{Path to grade file (\code{.rds}).}
}
\value{
Render the report for a grade (in \verb{<course>/reports/grades/}), save it to PNG, and update the \code{grade} file.
}
\description{
This function produces a .PNG (using \code{ggplot2})
}
