% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_rubric.R
\name{render_rubric}
\alias{render_rubric}
\title{Render rubric}
\usage{
render_rubric(course_id, assignment_id)
}
\arguments{
\item{course_id}{Course ID}

\item{assignment_id}{Assignment ID name}
}
\value{
Saves a PDF into the course's rubrics folder.
}
\description{
Render rubric
}
