% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_student.R
\name{render_student}
\alias{render_student}
\title{Render grade report for a student}
\usage{
render_student(
  course_id,
  goes_by,
  drop_lowest = NULL,
  gg_height = 10,
  to_file = FALSE
)
}
\arguments{
\item{course_id}{Course ID}

\item{goes_by}{Student name (goes_by column)}

\item{drop_lowest}{Optional; if there are \code{assignment_category}'s for which you want to drop the lowest grade for each student,
provide those categories here as a character vector.}

\item{gg_height}{Height of the \code{ggplot2} output, in inches.}

\item{to_file}{Boolean; if \code{TRUE}, the report will be saved to the class's folder structure (course > reports > students)}
}
\value{
A list.
}
\description{
Render grade report for a student
}
