% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_students.R
\name{render_students}
\alias{render_students}
\title{A wrapper for \code{render_student()}.}
\usage{
render_students(course_id, drop_lowest = NULL, gg_height = 10, to_file = TRUE)
}
\arguments{
\item{course_id}{Course ID}

\item{drop_lowest}{Optional; if there are \code{assignment_category}'s for which you want to drop the lowest grade for each student,
provide those categories here as a character vector.}

\item{gg_height}{Height of the \code{ggplot2} output, in inches.}

\item{to_file}{Boolean; if \code{TRUE}, the report will be saved to the class's folder structure (course > reports > students)}
}
\value{
desc
}
\description{
Loop through each student and print out a grade report for them.
}
