% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lateness-in-pivot.R
\name{replace_raw_with_lateness_scores}
\alias{replace_raw_with_lateness_scores}
\title{Replace Raw Scores with Lateness Scores
This function replaces original raw points with the scores after lateness
policies have been applied.}
\usage{
replace_raw_with_lateness_scores(wide_processed_data, lateness_scores)
}
\arguments{
\item{wide_processed_data}{A processed data in wide format}

\item{lateness_scores}{A dataframe with scores after lateness}
}
\value{
A data frame
}
\description{
Replace Raw Scores with Lateness Scores
This function replaces original raw points with the scores after lateness
policies have been applied.
}
