% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/review_google_form.R
\name{review_google_form}
\alias{review_google_form}
\title{Review reading assignments submitted via Google Form, by printing submissions to the Console}
\usage{
review_google_form(
  roster,
  form_url,
  form_cols = c("datetime", "email", "topic", "which", "question", "abstract"),
  week_of_class_start = 2
)
}
\arguments{
\item{roster}{The course roster of students, to be formatted the same as result of \code{view_students()}.}

\item{form_url}{The URL to the \code{GoogleSheet} containing \code{GoogleForm} responses.
Make sure link sharing is turned On. Student emails need to be collected with each submission.}

\item{form_cols}{A vector of new (simple) names to give the GoogleSheet columns. They can occur in any order needed to match your GoogleSheet, but
all the names listed in the default have to be assigned to a column in your form.
The column \code{topic} is used to sort submissions by submission topic/theme.
The column \code{which} is used to sort submissions within a topic/theme into groups before printing.
Currently two groups are recognized by the function: \code{"Abstract"} and \code{"Discussion Q&A"}.
If the \code{which} column contains the value \code{"Abstract"}, the function will look for a reading summary in a column named \code{abstract}.
If the \code{which} column contains the value \code{"Discussion Q&A"}, the function will look for a discussion question in a column named \code{question},
and it will look for a response in a column named \code{abstract}.}

\item{week_of_class_start}{Indicate which week of the year the class started in,
in order for the function to assign each timestamp submission to a week of the cours.}
}
\value{
Submitted reading assignments are printed to the Console, sorted by week and assignment type. At the end of the loop, summary tables are returned as output.
}
\description{
Review reading assignments submitted via Google Form, by printing submissions to the Console
}
