% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/review_timesheet.R
\name{review_timesheet}
\alias{review_timesheet}
\title{Review student timesheets that are submitted by Google Forms.}
\usage{
review_timesheet(
  roster,
  form_url,
  form_cols = c("email", "date", "hours", "percent"),
  week_of_class_start = 2
)
}
\arguments{
\item{form_url}{The URL to the \code{GoogleSheet} containing \code{GoogleForm} responses.
Make sure link sharing is turned On. Student emails need to be collected with each submission.}

\item{form_cols}{A vector of new (simple) names to give the GoogleSheet columns. They can occur in any order needed to match your GoogleSheet, but
all the names listed in the default have to be assigned to a column in your form.
The column \code{hours} refers to the duration of a work session (must be a numerical value).
the column \code{percent} refers to the student's best estimate of what percent of the work session was spent on productive, focused work.}

\item{week_of_class_start}{Indicate which week of the year the class started in,
in order for the function to assign each timestamp submission to a week of the cours.}

\item{course_id}{Course ID.}
}
\value{
A \code{data.frame} in which each row is hours worked in a week by a student.
}
\description{
Review student timesheets that are submitted by Google Forms.
}
