% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_course.R
\name{setup_course}
\alias{setup_course}
\title{Setup a subfolder for a course}
\usage{
setup_course(course_id)
}
\arguments{
\item{course_id}{Character vector of ID for your course, e.g., "ENST_320"}
}
\value{
desc
}
\description{
Creates the course subfolder within your working directory.
This function will not replace any existing folders or files.
}
