% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple_syllabus_data.R
\docType{data}
\name{simple_syllabus_data}
\alias{simple_syllabus_data}
\title{Simple Syllabus Data}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 5 rows and 20 columns.
}
\usage{
simple_syllabus_data
}
\description{
Sample grades data for a simple syllabus, shown in "Case Study: A Simple Syllabus" data
}
\examples{
simple_syllabus_data

# yaml:as_yaml(simple_syllabus_data)

}
\keyword{datasets}
