% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-grades.R
\name{sum_max_pts}
\alias{sum_max_pts}
\alias{mean_max_pts}
\title{Aggregation for Maximum Possible Points Functions}
\usage{
sum_max_pts(grades_mat, category, assignments)

mean_max_pts(grades_mat, category, assignments)
}
\arguments{
\item{grades_mat}{Matrix with assignments + associated columns for each category}

\item{category}{Category name}

\item{assignments}{Assignment names for this category}
}
\value{
A matrix with the additional computation of the maximum possible points for each category
}
\description{
A collection of functions to compute the maximum possible points for each category
\itemize{
\item \code{sum_max_pts()} computes the maximum possible points for each category as the sum of maximum points of each assignment within the category
\item \code{mean_max_pts()} computes the maximum possible points for each category as the average of maximum points of each assignment within the category
}
}
\concept{{Aggregation for Maximum Possible Points Functions}}
