% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-grades.R
\name{until}
\alias{until}
\alias{add}
\alias{between}
\alias{scale_by}
\alias{after}
\alias{set_to}
\title{Lateness Functions}
\usage{
until(grades_mat, late_policy, original_late_mat, assignments)

add(grades_mat, late_policy, original_late_mat, assignments)

between(grades_mat, late_policy, original_late_mat, assignments)

scale_by(grades_mat, late_policy, original_late_mat, assignments)

after(grades_mat, late_policy, original_late_mat, assignments)

set_to(grades_mat, late_policy, original_late_mat, assignments)
}
\arguments{
\item{grades_mat}{Matrix with assignments + associated cols for that category}

\item{late_policy}{Relevant threshold/scalar for lateness policy}

\item{original_late_mat}{Matrix that saves original lateness values}

\item{assignments}{Assignment names for this category}
}
\value{
A matrix
}
\description{
A collection of functions to apply lateness policies.
\itemize{
\item \code{until()} determines if lateness for \code{assignments} is less than or equal to \code{late_policy}.
\item \code{add()} adds \code{late_policy} to score if \code{assignments} is determined as late.
\item \code{between()} determines if lateness for \code{assignments} is between \code{late_policy}.
\item \code{scale_by()} scales score by \code{late_policy} if \code{assignments} is determined as late.
\item \code{after()} determines if lateness for \code{assignments} is more than or equal to \code{late_policy}.
\item \code{set_to()} sets score to \code{late_policy} if \code{assignments} is determined as late.
}
}
\concept{{Lateness Functions}}
