% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cat_table.R
\name{update_category}
\alias{update_category}
\title{Add or update category to policy list}
\usage{
update_category(
  policy_list,
  name = "Category",
  slipdays = 0,
  late_time1 = "00:00:00",
  late_time2 = "00:00:00",
  late_scale1 = 1,
  late_scale2 = 0,
  weight = 0,
  drops = 0,
  weight_method = c("equally", "by_points"),
  clobber = "None",
  assigns = c()
)
}
\arguments{
\item{policy_list}{a policy list if adding category to previous one.}

\item{name}{name of category.}

\item{slipdays}{allowed slip days.}

\item{late_time1}{cutoff for first lateness policy.}

\item{late_time2}{cutoff for second lateness policy.}

\item{late_scale1}{scale for first lateness policy.}

\item{late_scale2}{scale for second lateness policy.}

\item{weight}{weight for category.}

\item{drops}{allowed drops for category.}

\item{weight_method}{whether it's weighted equally or by points.}

\item{clobber}{whether it's clobbered.}

\item{assigns}{any assignments included in category.}
}
\value{
A list.
}
\description{
Add or update category to policy list
}
\examples{
update_category(name = "Labs")
update_category(name = "Quizzes")
}
