% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_roster.R
\name{update_roster}
\alias{update_roster}
\title{Update student roster for a course}
\usage{
update_roster(course_id, students_url, verbose = TRUE)
}
\arguments{
\item{course_id}{Course ID}

\item{students_url}{URL to a GoogleSheet (with sharing settings set to "Anyone with the Link") with the roster.
This spreadsheet must have these columns: \code{goes_by} (the name the student goes by),
\code{first} (first name), \code{last} (last name), \code{email} (full email address) (other columns are fine and will be kept).}

\item{verbose}{Boolean; print updates to console?}
}
\value{
A \code{students.rds} file saved to the course subfolder.
}
\description{
Update student roster for a course
}
