% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate-policy.R
\name{validate_policy}
\alias{validate_policy}
\title{Validate Policy File}
\usage{
validate_policy(policy, gs)
}
\arguments{
\item{policy}{YAML policy file}

\item{gs}{Gradescope data}
}
\description{
Flattens and validates policy file
}
