% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view_assignments.R
\name{view_assignments}
\alias{view_assignments}
\title{View a list of assignments & their status for a course}
\usage{
view_assignments(course_id)
}
\arguments{
\item{course_id}{Course ID.}
}
\value{
A dataframe with one row per assignment.
}
\description{
View a list of assignments & their status for a course
}
