% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view_grades.R
\name{view_grades}
\alias{view_grades}
\title{Quick access to all grades for a course}
\usage{
view_grades(course_id)
}
\arguments{
\item{course_id}{Course ID.}
}
\value{
A character vector of assignment \code{.rds} files for the course.
}
\description{
Quick access to all grades for a course
}
