% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view_grades_quick.R
\name{view_grades_quick}
\alias{view_grades_quick}
\title{Quick version of \code{view_grades()}}
\usage{
view_grades_quick(course_id, assignment_id = NULL)
}
\arguments{
\item{course_id}{Course ID.}

\item{assignment_id}{Optionally specify a single assignment.}
}
\value{
A \code{data.frame}.
}
\description{
This function returns the grades recorded for a given course (and optionally a specified assignment).
Unlike \code{view_grades()}, no details on the grades are returned. This allows the function to be very quick.
}
