% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view_missing.R
\name{view_missing}
\alias{view_missing}
\title{View missing grades}
\usage{
view_missing(course_id, mode = "complete", verbose = FALSE)
}
\arguments{
\item{course_id}{Course ID.}

\item{mode}{How much info to provide? The \code{"simple"} mode just provides the essential info; the \code{"complete"} mode returns all information.}

\item{verbose}{Print updates to console?}
}
\value{
A dataframe.
}
\description{
If a student does not have a grade for an assignment that was already due, it is a missing grade.
}
