% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view_missing_quick.R
\name{view_missing_quick}
\alias{view_missing_quick}
\title{View missing quick}
\usage{
view_missing_quick(course_id, assignment_id = NULL)
}
\arguments{
\item{course_id}{Course ID.}

\item{assignment_id}{Optionally specify a single assignment.}
}
\value{
A \code{data.frame}.
}
\description{
This function returns the missing (not-yet-submitted) assignments for a given course (and optionally a specified assignment).
Unlike \code{view_missing()}, no details on the assignments are returned. This allows the function to be very quick.
}
