% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view_status.R
\name{view_status}
\alias{view_status}
\title{View status of submissions, grading, and anything missing.}
\usage{
view_status(course_id, assignment_id = NULL, return_only = NULL)
}
\arguments{
\item{course_id}{Course ID.}

\item{assignment_id}{Optionally specify a single assignment.}
}
\value{
A dataframe of students, assignments, and their \code{graded} status (\code{TRUE} or \code{FALSE}).
}
\description{
View status of submissions, grading, and anything missing.
}
