% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view_students.R
\name{view_students}
\alias{view_students}
\title{Quick look at student roster for a course}
\usage{
view_students(course_id)
}
\arguments{
\item{course_id}{Course ID}
}
\value{
Returns a \code{data.frame} of the student roster.
}
\description{
Quick look at student roster for a course
}
