% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view_unshared.R
\name{view_unshared}
\alias{view_unshared}
\title{View which grades have not yet been emailed to students}
\usage{
view_unshared(course_id, mode = "simple", verbose = TRUE)
}
\arguments{
\item{course_id}{Course ID.}

\item{mode}{How much info to provide? The \code{"simple"} mode just provides the essential info; the \code{"complete"} mode returns all information.}

\item{verbose}{Print updates to console?}
}
\value{
A vector of grade files that have not yet been shared over email.
}
\description{
View which grades have not yet been emailed to students
}
