% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-grades.R
\name{calculate_grades}
\alias{calculate_grades}
\title{Calculate Grades
This function calculates all grades based on the policy file.}
\usage{
calculate_grades(gs, policy)
}
\arguments{
\item{gs}{A gradescope dataframe with students as rows and assignment information across the columns.}

\item{policy}{A  policy file}
}
\value{
A data frame
}
\description{
Calculate Grades
This function calculates all grades based on the policy file.
}
