% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-grades.R
\name{calculate_grades}
\alias{calculate_grades}
\title{Calculate Grades
This function calculates all grades based on the policy file.}
\usage{
calculate_grades(gs, policy)
}
\arguments{
\item{gs}{A Gradescope dataframes}

\item{policy}{A  policy file}
}
\value{
A Gradescope data frame with computed categories scores appended
}
\description{
Calculate Grades
This function calculates all grades based on the policy file.
}
