% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-gs.R
\name{check_data_format}
\alias{check_data_format}
\title{Check Formatting of Gradescope Data}
\usage{
check_data_format(gs, verbose = FALSE)
}
\arguments{
\item{gs}{Gradescope data frame}

\item{verbose}{Whether or not to print messages and warnings}
}
\value{
Same gs dataframe if no errors.
}
\description{
This functions checks the column names throughout the Gradescope data.
There must be an SID column and at least one assignment.
It also gives an alert for what ID columns and assignments are in the data.
}
\examples{
check_data_format(gs_demo, verbose = TRUE)

}
