% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/first_functions.R
\name{convert_to_min}
\alias{convert_to_min}
\title{Convert hms format to minutes using lubridate}
\usage{
convert_to_min(hms)
}
\arguments{
\item{hms}{A time in hms format}
}
\value{
A numeric vector.
}
\description{
Convert hms format to minutes using lubridate
}
\examples{
convert_to_min("11:11:11")
convert_to_min("22:00:00")
}
