% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-gs.R
\name{drop_ungraded_assignments}
\alias{drop_ungraded_assignments}
\title{Drop Ungraded Assignments}
\usage{
drop_ungraded_assignments(gs, verbose = FALSE)
}
\arguments{
\item{gs}{A Gradescope data frame}

\item{verbose}{whether or not to print message with the graded and ungraded assignments}
}
\value{
same dataframe without graded assignments
}
\description{
This functions drops any assignments that have no grades for any students and replaced -Inf values
}
