% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/email_grades.R
\name{email_grades}
\alias{email_grades}
\title{Send grades to students over email}
\usage{
email_grades(
  course_id,
  assignment_id = NULL,
  student_id = NULL,
  your_email = "ekezell@sewanee.edu",
  unshared_only = TRUE,
  json_path = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{course_id}{Course ID.}

\item{assignment_id}{Optionally specify an assignment; otherwise unsent grades from all assignments will be sent.}

\item{student_id}{Optionally specify a student; otherwise unsent grades for all students will be sent.}

\item{your_email}{Your email address.}

\item{json_path}{File path to your \code{json} credentials on your machine. This is necessary to send emails via the package \code{gmailr}.}

\item{verbose}{Print updates to console?}
}
\value{
desc
}
\description{
Send grades to students over email
}
