% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/email_grades.R
\name{email_grades}
\alias{email_grades}
\title{Send grades to students over email}
\usage{
email_grades(
  course_id,
  assignment_id = NULL,
  student_id = NULL,
  your_email = "ekezell@sewanee.edu",
  email_body =
    "Attached you will find your grade on this assignment.\\n\\nIf applicable, please recall that any overdue assignment (ones currently with a zero) can be updated if you submit the make-up materials.\\n\\nPlease let me know if you have any questions or concerns.\\n\\nBest wishes,\\nProf. Ezell",
  unshared_only = TRUE,
  json_path = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{course_id}{Course ID.}

\item{assignment_id}{Optionally specify an assignment; otherwise unsent grades from all assignments will be sent.}

\item{student_id}{Optionally specify a student; otherwise unsent grades for all students will be sent.}

\item{your_email}{Your email address.}

\item{email_body}{The main body of the email message; this will appear one line below the greeting to the student.}

\item{unshared_only}{Only share grades that have not yet been emailed? Default is \code{TRUE}.}

\item{json_path}{File path to your \code{json} credentials on your machine. This is necessary to send emails via the package \code{gmailr}.}

\item{verbose}{Print updates to console?}
}
\value{
desc
}
\description{
Send grades to students over email
}
