% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/email_test.R
\name{email_test}
\alias{email_test}
\title{Send a test email}
\usage{
email_test(
  recipient,
  your_email,
  intro = "Dear STUDENT,",
  signoff = "Best wishes,\\nProf. Ezell",
  email_body = "This is a test email using the gradebook R package.",
  json_path,
  verbose = TRUE
)
}
\arguments{
\item{recipient}{Receipient email address, for testing.}

\item{your_email}{Your email address.}

\item{json_path}{File path to your \code{json} credentials on your machine. This is necessary to send emails via the package \code{gmailr}.}

\item{verbose}{Print updates to console?}
}
\value{
An email is sent, updates are written to Console.
}
\description{
Send a test email
}
