% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-grades.R
\name{equally_weighted}
\alias{equally_weighted}
\alias{weighted_by_points}
\alias{max_score}
\alias{min_score}
\alias{weighted_mean}
\alias{none}
\title{Aggregation Functions}
\usage{
equally_weighted(grades_mat, category, assignments, weights = c())

weighted_by_points(grades_mat, category, assignments, weights = c())

max_score(grades_mat, category, assignments, weights = c())

min_score(grades_mat, category, assignments, weights = c())

weighted_mean(grades_mat, category, assignments, weights)

none(grades_mat, category, assignments, weights = c())
}
\arguments{
\item{grades_mat}{Matrix with assignments + associated cols for that category}

\item{category}{Category name}

\item{assignments}{Assignment names for this category}

\item{weights}{Weights for \code{weighted_mean}, default to NULL with other aggregation method}
}
\value{
A matrix with the additional computation of the aggregated category score.
}
\description{
A collection of functions to compute aggregated scores for each category.
\itemize{
\item \code{equally_weighted()} computes the aggregated category score by taking the mean of assignment scores.
\item \code{weighted_by_points()} computes the aggregated category score by taking the weighted mean of assignment scores using a weights vector.
\item \code{max_score()} computes the aggregated category score by taking the maximum of assignment scores.
\item \code{min_score()} computes the aggregated category score by taking the minimum of assignment scores.
\item \code{none()} is to be used if there is only 1 score, returns it; otherwise defaults to \code{equally_weighted}
}
}
\concept{{Aggregation Functions}}
