% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate-grades.R
\name{equally_weighted}
\alias{equally_weighted}
\alias{weighted_by_points}
\alias{max_score}
\alias{min_score}
\alias{weighted_mean}
\alias{none}
\title{Aggregation Functions}
\usage{
equally_weighted(grades_mat, category, assignments, weights = c())

weighted_by_points(grades_mat, category, assignments, weights = c())

max_score(grades_mat, category, assignments, weights = c())

min_score(grades_mat, category, assignments, weights = c())

weighted_mean(grades_mat, category, assignments, weights)

none(grades_mat, category, assignments, weights = c())
}
\arguments{
\item{grades_mat}{Matrix with assignments + associated cols for that category}

\item{category}{Category name}

\item{assignments}{Assignment names for this category}

\item{weights}{Weights for \code{weighted_mean}, default to NULL with other aggregation method}
}
\value{
A matrix
}
\description{
A collection of functions to compute scores for categories.
\itemize{
\item \code{equally_weighted()} computes a mean.
\item \code{weighted_by_points()} computes a weighted mean using a weights vector.
\item \code{max_score()} computes the max score.
\item \code{min_score()} computes the min score after optionally dropping lowest scores.
\item \code{none()} if there is only 1 score, returns it, otherwise defaults to \code{equally_weighted}
}
}
\concept{{Aggregation Functions}}
