% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate-policy.R
\name{flatten_policy}
\alias{flatten_policy}
\title{Reshape policy file from nested to flat}
\usage{
flatten_policy(policy)
}
\arguments{
\item{policy}{A valid policy file stored as a list.}
}
\value{
A single level list of all categories and subcategories ordered where
all leaves will precede the parent category in the list order.
}
\description{
First propagates \code{lateness} to all child categories then cycles through the
top-level categories of a policy file and unnests all subcategories to create
a single level list of all categories and subcategories.
}
\examples{
# Example
flatten_policy(policy_demo)
}
