% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-gs.R
\name{get_assignments}
\alias{get_assignments}
\title{Get the Assignment Names for Gradescope Data}
\usage{
get_assignments(gs, verbose = FALSE)
}
\arguments{
\item{gs}{Gradescope dataframe}

\item{verbose}{Whether or not to print assignment names}
}
\value{
A vector of the names of the assignments in the dataframe
}
\description{
This function identifies the assignments from Gradescope data and returns the assignments' names.
}
\examples{
get_assignments(gs_demo, verbose = TRUE)

}
