% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-gs.R
\name{get_assignments}
\alias{get_assignments}
\title{Get the Assignment Names for Gradescope Data}
\usage{
get_assignments(gs, verbose = FALSE)
}
\arguments{
\item{gs}{unprocessed Gradescope dataframe}

\item{verbose}{whether or not to print assignment names}
}
\value{
vector
}
\description{
This function identified the assignments from Gradescope data
}
