% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-grades.R
\name{get_category_grade}
\alias{get_category_grade}
\title{Calculate Single Category Grade
This function calculates the grade of a single category based on the policy file.}
\usage{
get_category_grade(grades_mat, policy_item)
}
\arguments{
\item{grades_mat}{Matrix with assignments + associated cols for that category}

\item{policy_item}{An item from policy file}
}
\value{
A data frame
}
\description{
Calculate Single Category Grade
This function calculates the grade of a single category based on the policy file.
}
