% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-grades.R
\name{get_grades}
\alias{get_grades}
\title{Get Grades
This function processes the Gradescope data and the policy file, and then calculates all grades based on the policy file.}
\usage{
get_grades(gs, policy, verbose = FALSE)
}
\arguments{
\item{gs}{A gradescope dataframe with students as rows and assignment information across the columns.}

\item{policy}{A policy file}

\item{verbose}{if FALSE, throws error if no assignments found in gs}
}
\value{
A data frame
}
\description{
Get Grades
This function processes the Gradescope data and the policy file, and then calculates all grades based on the policy file.
}
