% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-grades.R
\name{get_grades}
\alias{get_grades}
\title{Get Grades}
\usage{
get_grades(gs, policy, verbose = FALSE)
}
\arguments{
\item{gs}{A Gradescope dataframe}

\item{policy}{A grading policy file as an R list}

\item{verbose}{if FALSE, throws error if no assignments found in gs}
}
\value{
A dataframe of the original Gradescope data with computed categories' scores appended as additional columns
}
\description{
This function processes the Gradescope data and the policy file,
and then calculates all grades based on the criteria from the policy file and
the assignment scores from the Gradescope data. This function assumes the
YAML policy file and the Gradescope .csv file were read in using the
\code{read_policy()} and \code{read_gs()} functions respectively, which check for proper
data format.
}
\examples{
get_grades(gs = gs_demo, policy = policy_demo, verbose = TRUE)

}
