% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate-grades.R
\name{get_grades}
\alias{get_grades}
\title{Calculate Grades
This function calculates all grades based on the policy file.}
\usage{
get_grades(gs, policy)
}
\arguments{
\item{gs}{A processed data in wide format}

\item{policy}{A policy file}
}
\value{
A data frame
}
\description{
Calculate Grades
This function calculates all grades based on the policy file.
}
