% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-gs.R
\name{get_id_cols}
\alias{get_id_cols}
\title{Get the ID Columns for Gradescope Data}
\usage{
get_id_cols(gs, verbose = FALSE)
}
\arguments{
\item{gs}{Gradescope dataframe}

\item{verbose}{Whether or not to return an alert of assignments}
}
\value{
A vector of the names of the ID columns in the dataframe
}
\description{
This function identifies the ID columns from Gradescope data.
}
