% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-gs.R
\name{get_id_cols}
\alias{get_id_cols}
\title{Get the ID Columns for Gradescope Data}
\usage{
get_id_cols(gs, verbose = FALSE)
}
\arguments{
\item{gs}{Gradescope dataframe}

\item{verbose}{whether or not to return an alert of assignments}
}
\value{
a list of id columns
}
\description{
This function identified the id columns from gradescope data
}
