% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gs_demo.R
\docType{data}
\name{gs_demo}
\alias{gs_demo}
\title{Gradescope Demo Data}
\format{
\subsection{\code{gs_demo}}{

A data frame with 10 rows and 44 columns:
\describe{
\item{Names}{Student name}
\item{Email}{Student email}
\item{SID}{Student ID number}
\item{Sections}{Student class section}
\item{Lab 1}{An assignment called Lab 1}
\item{Lab 1 - Max Points}{Maximum possible points for Lab 1}
\item{Lab 1 - Lateness (H:M:S)}{How late Lab 1 was turned in, formatted HH:MM:SS}
\item{Lab 1 - Submission Time}{When Lab 1 was turned in, formatted MM/DD/YY HH:MM}
...
}
}
}
\usage{
gs_demo
}
\description{
An example dataset that mimics Gradescope format with students as rows
and assignment information across the columns.
All assignments follow the same four-column format of Lab 1, as noted below.
Assignments in \code{gs_demo} are Lab 1, Lab 2, Lab 3, Lab 4, Lab 5, Lab 6, Discussion 1,
Discussion 2, Midterm and Final.
Any other columns that do not follow this format are designated as ID columns.
}
\examples{
gs_demo

}
\keyword{datasets}
