% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gs_demo.R
\docType{data}
\name{gs_demo}
\alias{gs_demo}
\title{TBD}
\format{
\subsection{\code{gs_demo}}{

TBD
\describe{
\item{country}{Country name}
\item{iso2, iso3}{2 & 3 letter ISO country codes}
\item{year}{Year}
...
}
}
}
\usage{
gs_demo
}
\description{
TBD
}
\keyword{datasets}
