% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate-grades.R
\name{mean_lateness}
\alias{mean_lateness}
\alias{sum_lateness}
\alias{max_lateness}
\title{Aggregation for Lateness Functions}
\usage{
mean_lateness(grades_mat, category, assignments)

sum_lateness(grades_mat, category, assignments)

max_lateness(grades_mat, category, assignments)
}
\arguments{
\item{grades_mat}{Matrix with assignments + associated cols for that category}

\item{category}{Category name}

\item{assignments}{Assignment names for this category}
}
\value{
A matrix
}
\description{
A collection of functions to computes lateness for category.
\itemize{
\item \code{mean_lateness()} computes a sum.
\item \code{sum_lateness()} computes a mean.
\item \code{max_lateness()} computes a max.
}
}
\concept{{Aggregation for Lateness Functions}}
