% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/policy_demo.R
\docType{data}
\name{policy_demo}
\alias{policy_demo}
\title{Policy Demo File}
\format{
\subsection{\code{policy_demo}}{

A demo policy file
\describe{
\description{Coursewide: includes class info}
\item{course_name}{course name}
\item{description}{any relevant course description, purely for the user}
\description{Categories: each category in syllabus}
\item{name}{category name}
\item{slipdays}{number of slip days}
...
\description{Cutoff: grade cutoffs}
\item{Cutoff}{each cutoff reflects the lower bound for each letter grade: A,B,C,D,F}
}
}
}
\usage{
policy_demo
}
\description{
A fake policy file that mimics a hypothetical class syllabus
}
\keyword{datasets}
