% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/policy_demo.R
\docType{data}
\name{policy_demo}
\alias{policy_demo}
\title{Policy Demo File}
\format{
\subsection{\code{policy_demo}}{

A demo policy file
\describe{
\item{TBD}{student name}
\description{TBD}
...
}
}
}
\usage{
policy_demo
}
\description{
A fake policy file that mimics a hypothetical class syllabus
}
\keyword{datasets}
