% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process-gs.R
\name{process_gs}
\alias{process_gs}
\title{Process Gradescope .csv}
\usage{
process_gs(gs, drop_ungraded = TRUE, verbose = FALSE)
}
\arguments{
\item{gs}{A gradescope dataframe with students as rows and assignment information across the columns.}

\item{drop_ungraded}{whether or not to drop ungraded assignments}

\item{verbose}{whether or not to print messages}
}
\value{
dataframe
}
\description{
This function processes the gradescope data before graidng.  Ungraded assignments are dropped, and
eventually, this will also merge duplicated students and deal with students with no SID.
}
