% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process-gs.R
\name{process_gs}
\alias{process_gs}
\title{Process Gradescope Data}
\usage{
process_gs(gs_data, wide_format = FALSE)
}
\arguments{
\item{gs_data}{A dataframe (csv from Gradescope)}

\item{wide_format}{whether or not to pivot into wide format}
}
\value{
A dataframe
}
\description{
This functino merges any duplicates students and drops ungraded assignments
for faster processing.
}
