% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process-policy.R
\name{process_policy}
\alias{process_policy}
\title{Validate Policy File}
\usage{
process_policy(policy, verbose = FALSE)
}
\arguments{
\item{policy}{YAML policy file}

\item{verbose}{if FALSE, throws error if no assignments found in gs}
}
\description{
Flattens and validates policy file
}
