% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-grades.R
\name{raw_over_max}
\alias{raw_over_max}
\title{Score Functions}
\usage{
raw_over_max(grades_mat, assignments)
}
\arguments{
\item{grades_mat}{Matrix with assignments + associated columns for that category}

\item{assignments}{Assignment names for this category}
}
\value{
A matrix with computed assignment scores
}
\description{
A collection of functions to calculate the percentage score and save back into
the column with the assignment score
\itemize{
\item \code{raw_over_max()} computes score by dividing the original assignment score by the maximum possible points for the assignment
}
}
\concept{{Score Functions}}
