% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate-grades.R
\name{raw_over_max}
\alias{raw_over_max}
\title{Score Functions}
\usage{
raw_over_max(grades_mat, assignments)
}
\arguments{
\item{grades_mat}{Matrix with assignments + associated cols for that category}

\item{assignments}{Assignment names for this category}
}
\value{
a matrix
}
\description{
\itemize{
\item \code{raw_over_max()} computes score by dividing raw points by max points
}
}
\concept{{Score functions}
A collection of functions that calculate score for assignments}
