% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process-gs.R
\name{read_gs}
\alias{read_gs}
\title{Read Gradescope .csv}
\usage{
read_gs(path)
}
\arguments{
\item{path}{Path to Gradescope CSV}
}
\value{
dataframe
}
\description{
This functions reads the Gradescope .csv, converts NA values to zeros and
computes scores for each assignment
}
