% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-gs.R
\name{read_gs}
\alias{read_gs}
\title{Read Gradescope .csv}
\usage{
read_gs(path, verbose = FALSE)
}
\arguments{
\item{path}{Path to Gradescope CSV}

\item{verbose}{whether or not to print messages}
}
\value{
dataframe
}
\description{
This functions reads the Gradescope .csv, checks for correct format.

Now it is favored to use read_files.
This function is left for backwards-compatibility.
}
