% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate-gs.R
\name{read_gs}
\alias{read_gs}
\title{Read Gradescope .csv}
\usage{
read_gs(path)
}
\arguments{
\item{path}{Path to Gradescope CSV}
}
\value{
dataframe of Gradescope csv
}
\description{
This functions reads the Gradescope .csv and processes certain columns into
the expected datatypes.
}
