% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-gs.R
\name{read_gs}
\alias{read_gs}
\title{Read Gradescope .csv}
\usage{
read_gs(path, drop_ungraded = TRUE, verbose = FALSE)
}
\arguments{
\item{path}{Path to Gradescope CSV}

\item{drop_ungraded}{whether or not to drop ungraded assignments}

\item{verbose}{whether or not to print messages}
}
\value{
dataframe
}
\description{
This functions reads the Gradescope .csv, checks for correct format,
converts NA values to zeros and computes scores for each assignment.
This function can also drop ungraded assignments
}
