% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-policy.R
\name{read_policy}
\alias{read_policy}
\title{Read Grading Policy .yml}
\usage{
read_policy(path, verbose = FALSE)
}
\arguments{
\item{path}{Path to YAML policy file}

\item{verbose}{Whether or not to print messages}
}
\value{
R list of the policy file, if no errors
}
\description{
This functions reads the grading policy YAML file and checks for correct format
}
