% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process-policy.R
\name{reconcile_policy_with_gs}
\alias{reconcile_policy_with_gs}
\title{Reconcile policy file with Gradescope data}
\usage{
reconcile_policy_with_gs(policy, gs, verbose = FALSE)
}
\arguments{
\item{policy}{A valid policy file stored as a list.}

\item{gs}{Gradescope data}

\item{verbose}{if FALSE, throws error if no assignments found in gs}
}
\value{
A policy list
}
\description{
This function drops any assignments present in the policy file that are not in the Gradescope data.
}
