% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_grades.R
\name{render_grades}
\alias{render_grades}
\title{Render reports for a bunch of grades
This function is a wrapper for \code{grade_render()}.}
\usage{
render_grades(course_id, assignment_id = NULL)
}
\arguments{
\item{course_id}{Course ID.}

\item{assignment_id}{Assignment ID.}
}
\value{
Updates grade file and grade report for all matching grades.
}
\description{
Render reports for a bunch of grades
This function is a wrapper for \code{grade_render()}.
}
