% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_student.R
\name{render_student}
\alias{render_student}
\title{Render grade report for a student}
\usage{
render_student(course_id, goes_by)
}
\arguments{
\item{course_id}{Course ID}

\item{goes_by}{Student name (goes_by column)}
}
\value{
A list.
}
\description{
Render grade report for a student
}
