% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-grades.R
\name{score}
\alias{score}
\alias{aggregation}
\alias{lateness}
\alias{drop_n_lowest}
\alias{aggregation_max_pts}
\alias{aggregation_lateness}
\title{Key Functions}
\usage{
score(grades_mat, policy_line, category, assignments, weights = c())

aggregation(grades_mat, policy_line, category, assignments, weights = c())

lateness(grades_mat, policy_line, category, assignments, weights = c())

drop_n_lowest(grades_mat, policy_line, category, assignments, weights = c())

aggregation_max_pts(
  grades_mat,
  policy_line,
  category,
  assignments,
  weights = c()
)

aggregation_lateness(
  grades_mat,
  policy_line,
  category,
  assignments,
  weights = c()
)
}
\arguments{
\item{grades_mat}{Matrix with assignments + associated cols for that category}

\item{policy_line}{Policy list item for that key}

\item{category}{Category name}

\item{assignments}{Assignment names for this category}

\item{weights}{Weights for \code{weighted_mean}}
}
\value{
A matrix
}
\description{
A collection of functions to refer to correct classification of grading functions using policy file keys.
\itemize{
\item \code{score()} computes the percentage score and saves back into raw points col
\item \code{aggregation()} computes score for category
\item \code{lateness()} applies any relevant lateness penalty.
\item \code{drop_n_lowest()} drops n lowest assignment scores; if n >= num of assignments, return top assignment
\item \code{aggregation_max_pts()} computes max points for category.
\item \code{aggregation_lateness()} computes max points for category.
}
}
\concept{{Key Functions}}
