% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate-grades.R
\name{score}
\alias{score}
\alias{aggregation}
\alias{lateness}
\alias{drops}
\alias{aggregation_max_pts}
\title{Key Functions}
\usage{
score(grades_mat, policy_line, category, assignments)

aggregation(grades_mat, policy_line, category, assignments)

lateness(grades_mat, policy_line, category, assignments)

drops(grades_mat, policy_line, category, assignments)

aggregation_max_pts(grades_mat, policy_line, category, assignments)
}
\arguments{
\item{grades_mat}{Matrix with assignments + associated cols for that category}

\item{policy_line}{Policy list item for that key}

\item{category}{Category name}

\item{assignments}{Assignment names for this category}
}
\value{
A matrix
}
\description{
A collection of functions to refer to correct classification of grading functions using policy file keys.
\itemize{
\item \code{score()} computes the percentage score and saves back into raw points col
\item \code{aggregation()} computes score for category
\item \code{lateness()} applies any relevant lateness penalty.
\item \code{drops()} drops n lowest/highest assignment scores; note that this is NOT implemented yet.
\item \code{aggregation_max_pts()} computes max points for category.
}
}
\concept{{Key Functions}}
