% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate-grades.R
\name{sum_max_pts}
\alias{sum_max_pts}
\alias{mean_max_pts}
\title{Aggregation for Max Points Functions}
\usage{
sum_max_pts(grades_mat, category, assignments)

mean_max_pts(grades_mat, category, assignments)
}
\arguments{
\item{grades_mat}{Matrix with assignments + associated cols for that category}

\item{category}{Category name}

\item{assignments}{Assignment names for this category}
}
\value{
A matrix
}
\description{
A collection of functions to computes max points for category.
\itemize{
\item \code{sum_max_pts()} computes a sum.
\item \code{mean_max_pts()} computes a mean.
}
}
\concept{{Aggregation for Max Points Functions}}
