% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process-policy.R
\name{validate_policy}
\alias{validate_policy}
\title{Validate Policy File}
\usage{
validate_policy(policy, gs, quiet = FALSE)
}
\arguments{
\item{policy}{YAML policy file}

\item{gs}{Gradescope data}

\item{quiet}{if FALSE, throws error if no assignments found in gs}
}
\description{
Flattens and validates policy file
}
