% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view_missing.R
\name{view_missing}
\alias{view_missing}
\title{View missing grades}
\usage{
view_missing(course_id)
}
\arguments{
\item{course_id}{Course ID.}
}
\value{
A dataframe.
}
\description{
If a student does not have a grade for an assignment that was already due, it is a missing grade.
}
